Plot BigWig Coverage for Genes of Interest by a Given Variable

Plot BigWig coverage for genes of interest colored by a given variable

Usage

plot_gene_coverage_by_var(
  genes_of_interest = "NRL",
  cell_metadata,
  bigwig_tbl,
  group_by = "batch",
  values_of_interest = NULL,
  organism = c("human", "mouse"),
  edb = NULL,
  heights = c(3, 1),
  scale_y = "log10",
  reverse_x = FALSE,
  start = NULL,
  end = NULL,
  summarize_transcripts = FALSE,
  ...
)

Arguments

genes_of_interest

Gene of interest

cell_metadata

a dataframe with cell metadata from object

bigwig_tbl

a tibble with colnames "name", "bigWig", and "sample_id" matching the filename, absolute path, and sample name of each cell in the cell_metadata

group_by

Variable to color by

values_of_interest

values of interest

organism

human (default) or mouse

edb

ensembldb object

heights

The heights of each row in the grid of plot

scale_y

whether to scale coverage

reverse_x

whether to reverse x axis

start

start coordinates

end

end coordinates

summarize_transcripts

whether to summarize transcript counts

...

extra arguments passed to plotCoverageFromEnsembldb

Value

a ggplot with coverage faceted by group_by